+++ tests/testthat/test-get_attribution.R
@@ -4,14 +4,21 @@ test_that("get_attribution works", {
   uuid <- get_uuid(name = "Acropora cervicornis")
   # Expect true
   expect_true(is.list(get_attribution(uuid = uuid)))
-  expect_true(is.null(get_attribution(uuid = uuid, text = TRUE)))
   # Expect equal
   uuid <- get_uuid(name = "Scleractinia", n = 5)
-  expect_equal(length(get_attribution(uuid = uuid)), 5)
+  ## multiple uuids
+  expect_equal(length(get_attribution(uuid = uuid)), 1)
   expect_message(get_attribution(uuid = uuid, text = TRUE))
   # Check img arg
   img <- get_phylopic(uuid = uuid[1])
   expect_true(is.list(get_attribution(img = img)))
+  # Check permalink generation
+  perm <- get_attribution(uuid = uuid, permalink = TRUE)
+  expect_true("permalink" %in% names(perm))
+  expect_message(get_attribution(uuid = uuid, text = TRUE, permalink = TRUE))
+  ## one uuid handling
+  expect_equal(length(get_attribution(uuid = uuid[1], 
+                                      text = TRUE, permalink = TRUE)), 3)
 
   # Expect error
   expect_error(get_attribution(uuid = NULL))