+++ NEWS.md
@@ -1,6 +1,7 @@
 # rphylopic (development version)
 
 * Added add_phylopic_legend (#83)
+* Added permalink generation option to get_attribution (#81)
 
 # rphylopic 1.3.0